refactor: reorganizar proyecto con estructura src/ y tests/
