+++ src/__init__.py
@@ -1,0 +1,7 @@
+"""Paquete calculadora con GUI y CLI.
+
+Módulos:
+    calculator: Funciones matemáticas básicas
+    gui: Interfaz gráfica con tkinter
+    cli: Interfaz de línea de comandos
+"""
