+++ src/cli.py
@@ -1,11 +1,10 @@
-#!/usr/bin/env python3
 """Archivo principal de ejemplo para el proyecto Team Practice.
 
 Demuestra cómo usar los módulos del proyecto y cómo estructurar
 un punto de entrada principal.
 """
 
-from calculator import add, subtract, multiply, divide, power, valor_maximo, valor_minimo, abs_value
+from .calculator import add, subtract, multiply, divide, power, valor_maximo, valor_minimo, abs_value
 
 def main():
     """Función principal del programa."""
