+++ tests/__init__.py
@@ -1,0 +1,4 @@
+"""Suite de tests para Team Practice Calculator.
+
+Tests unitarios para los módulos calculator y gui.
+"""
